Revert "new reward function and short queue results (#33)"
This reverts commit 9615c7bb1a98ca71d44e8d7ab4dd2520f7c2787c.
