+++ stable_radical_optimization/initialize.py
@@ -34,10 +34,7 @@ with psycopg2.connect(**config.dbparams) as conn:
             atom_type varchar(2),
             buried_vol real,
             max_spin real,
-            atom_index int,
-            ie real,
-            ea real,
-            bde real
+            atom_index int
             );
             
         CREATE TABLE IF NOT EXISTS {table}_replay (
