+++ stable_radical_optimization/stable_rad_config.py
@@ -2,10 +2,10 @@ import os
 import alphazero.config as config
 
 # DB table names modified by the user according to their wish
-config.sql_basename = "StablePSJ_newreward"
+config.sql_basename = "StablePSJ"
 
 # Experiment id
-config.experiment_id = "new_reward"
+config.experiment_id = "fixing_node_visits_1"
 
 with open('/projects/rlmolecule/rlops_pass', 'r') as f:
     passwd = f.read().strip()
@@ -22,11 +22,9 @@ config.dbparams = {
 config.checkpoint_filepath = os.path.expandvars(
     f'/scratch/$USER/policy_checkpoints/{config.sql_basename}/{config.experiment_id}')
 
-# config.dirichlet_alpha = 0.5
 config.dirichlet_x = 0.5  # percentage to favor dirichlet noise vs. prior estimation. Smaller means less noise
 
 config.build_kwargs.update({'atom_additions':  ('C', 'N', 'O', 'S'),
-                            'sa_score_threshold': 3.5})
+                            'sa_score_threshold': 4.})
 
 config.max_atoms = 15
-config.min_reward = 0
