+++ stable_radical_optimization/submit_combined.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #SBATCH --account=rlmolecule
 #SBATCH --time=4:00:00
-#SBATCH --job-name az_stability_newreward_short
+#SBATCH --job-name az_stability
 # --- Policy Trainer ---
 #SBATCH --nodes=1
 #SBATCH --gres=gpu:2
